+++ tests/initialize_db.py
@@ -20,48 +20,18 @@ def main(db, csv):
         for k, v in db_config.items():
             os.environ['PG' + k.upper()] = str(v) if v else ""
 
-    if db_config.get('type', 'postgres') == 'mssql':
-        del db_config['type']
-        init_mssql(db_config, csv)
-    else:
-        init(db, csv)
+    init(db_config, csv)
 
 
-def init(db, csv):
-    # We'll shell out to `psql`, so set the environment variables for it:
-    # And create the table from the csv file with psql
-    with open(db) as f:
-        for k, v in yaml.load(f).items():
-            os.environ['PG' + k.upper()] = str(v) if v else ""
-
-    system("""psql -c "CREATE SCHEMA IF NOT EXISTS dedupe;" """)
-    system("""psql -c "DROP TABLE IF EXISTS dedupe.entries;" """)
-
-    system("""psql -c "
-         CREATE TABLE "dedupe.entries" (
-             uuid UUID,
-             first_name VARCHAR,
-             last_name VARCHAR,
-             ssn VARCHAR(11),
-             sex VARCHAR(1),
-             dob VARCHAR(10),
-             race VARCHAR,
-             ethnicity VARCHAR
-         );" """)
-    system("""psql -c "\copy dedupe.entries FROM '{}' WITH CSV HEADER;" """.format(csv))
-    system("""psql -c "alter table dedupe.entries add column entry_id SERIAL PRIMARY KEY;" """)
-    system("""psql -c "alter table dedupe.entries add column full_name VARCHAR;" """)
-    system("""psql -c "update dedupe.entries set full_name = first_name || ' ' || last_name;" """)
-
-
-def init_mssql(db, csv_file):
+def init(db, csv_file):
     # Connect to database and use cursor to insert csv data to table
     con = pymssql.connect(**db)
 
     # Create DB for testing. Do not nned to install sql-tools in Travis
     con.autocommit(True)
     c = con.cursor()
-    c.execute("IF NOT EXISTS(select * from sys.databases where name='test') CREATE DATABASE test")
+    c.execute(
+        "IF NOT EXISTS(select * from sys.databases where name='test') CREATE DATABASE test")
     c.close()
     con.close()
 
@@ -73,7 +43,8 @@ def init_mssql(db, csv_file):
             EXEC('CREATE SCHEMA dedupe')
             END""")
 
-    c.execute("IF OBJECT_ID('dedupe.entries', 'U') IS NOT NULL DROP TABLE dedupe.entries;")
+    c.execute(
+        "IF OBJECT_ID('dedupe.entries', 'U') IS NOT NULL DROP TABLE dedupe.entries;")
 
     # Lazy definition for the table; Can be made more precise
     c.execute("""CREATE TABLE dedupe.entries (
@@ -99,10 +70,13 @@ def init_mssql(db, csv_file):
             c.execute(query, tuple(data))
             con.commit()
 
-    c.execute("""ALTER TABLE dedupe.entries ADD entry_id INT IDENTITY(1,1) PRIMARY KEY""")
+    c.execute(
+        """ALTER TABLE dedupe.entries ADD entry_id INT IDENTITY(1,1) PRIMARY KEY""")
     c.execute("""ALTER TABLE dedupe.entries ADD full_name VARCHAR(100)""")
-    c.execute("""UPDATE dedupe.entries SET full_name = first_name + ' ' + last_name""")
-    c.execute("""UPDATE dedupe.entries SET first_name = NULL where first_name = ''""")
+    c.execute(
+        """UPDATE dedupe.entries SET full_name = first_name + ' ' + last_name""")
+    c.execute(
+        """UPDATE dedupe.entries SET first_name = NULL where first_name = ''""")
     c.execute("""UPDATE dedupe.entries SET last_name = NULL where last_name=''""")
     c.execute("""UPDATE dedupe.entries SET ssn = NULL where ssn=''""")
     c.execute("""UPDATE dedupe.entries SET sex = NULL where sex=''""")
