+++ tests/test_integration.py
@@ -1,30 +1,47 @@
 # Very rudimentary integration testing
-from ssdedupe import cli
+import json
+
 from click.testing import CliRunner
+
+from ssdedupe import cli
+
+import pymssql
+
 import tests.generate_fake_dataset as gen
 import tests.initialize_db as initdb
 
-import os
-import yaml
-import psycopg2 as psy
-import testing.postgresql
-
 
 def test_integration():
-    psql = testing.postgresql.Postgresql()
-    with open('db.yaml','w') as f:
-        yaml.dump(psql.dsn(), f)
+    # SQL Server credential for test on docker
+    db = {
+        'host': 'localhost',
+        'password': '1@34dedupe',
+        'user': 'sa',
+        'port': 15433
+    }
 
     pop = gen.create_population(2000)
     gen.create_csv(pop, 'pop.csv')
 
-    initdb.init('db.yaml', 'pop.csv')
-    
+    initdb.init(db, 'pop.csv')
+
+    with open('db.json', 'w') as f:
+        db['database'] = 'test'
+        json.dump(db, f)
+
     runner = CliRunner()
-    result = runner.invoke(cli.main, ['--config', 'config.yaml', '--db', 'db.yaml'])
+    result = runner.invoke(
+        cli.main, ['--config', 'config.yaml', '--db', 'db.json'])
     assert result.exit_code == 0
 
-    con = psy.connect(**psql.dsn())
+    con = pymssql.connect(**db)
     c = con.cursor()
     c.execute("SELECT count(distinct dedupe_id) FROM dedupe.entries")
-    assert(c.fetchone()[0] < 4000) # Rudimentary quality check
+    assert(c.fetchone()[0] < 4000)
+    # Rudimentary quality check
+
+    # Delete database test created
+    con.autocommit(True)
+    c.execute("USE master")
+    c.execute("DROP DATABASE test")
+    con.close()
