+++ tests/test_ssdedupe.py
@@ -30,6 +30,8 @@ def test_content(response):
     """
     # from bs4 import BeautifulSoup
     # assert 'GitHub' in BeautifulSoup(response.content).title.string
+
+
 def test_command_line_interface():
     runner = CliRunner()
     # result = runner.invoke(cli.main)
